tutum-agent uninstalls dokku-alt
Tutum is a service to manage nodes with Docker inside them. Checkout the website on [tutum.co]

tutum-agent, which is the software that must be installed on each node to be managed by the Tutum service includes its own version of Docker embedded in the package (docker 1.5).

Installing both Dokku and Tutum on the same machine will result in one being deleted because the package tutum-agent replaces lxc-docker and Dokku depends on it, so it is deleted too.

I propose using tutum-agent as an alternative to lxc-docker.